Revert "Merge pull request #510 from VinceMacBuche/bug_4831/table_fromin_to_bigint"
This reverts commit 75c942b4c8dcb774484d83449005ec66f5a43eeb, reversing
changes made to 4cab352e01eb64da02b472c0fda932f0da44a25a.

+++ rudder-core/src/main/resources/reportsSchema.sql
@@ -78,7 +78,7 @@ create index expectedReportsNodes_versionId on expectedReportsNodes (nodeJoinKey
 create sequence serial START 101;
 
 CREATE TABLE RudderSysEvents (
-id bigint PRIMARY KEY default nextval('serial'),
+id integer PRIMARY KEY default nextval('serial'),
 executionDate timestamp with time zone NOT NULL, 
 nodeId text NOT NULL CHECK (nodeId <> ''),
 directiveId text NOT NULL CHECK (directiveId <> ''),
@@ -102,7 +102,7 @@ CREATE INDEX ruleId_idx on RudderSysEvents (ruleId);
 
 
 CREATE TABLE ArchivedRudderSysEvents (
-id bigint PRIMARY KEY,
+id integer PRIMARY KEY,
 executionDate timestamp with time zone NOT NULL, 
 nodeId text NOT NULL CHECK (nodeId <> ''),
 directiveId text NOT NULL CHECK (directiveId <> ''),
