+++ rudder-core/src/main/scala/com/normation/rudder/repository/RudderPropertiesRepository.scala
@@ -41,11 +41,11 @@ trait RudderPropertiesRepository {
   /**
    * Get the last report id processed by the non compliant report Logger.
    */
-  def getReportLoggerLastId: Box[Long]
+  def getReportLoggerLastId: Box[Int]
 
   /**
    * Update or create (if needed the last id processed by the non compliant report logger
    */
-  def updateReportLoggerLastId(newId: Long) : Box[Long]
+  def updateReportLoggerLastId(newId: Int) : Box[Int]
 
 }
